Address all code review findings from Phase 2 review
Important Improvements (§3.1-3.4):

§3.1 RegionCache: Emergency eviction null safety
- Filter nulls before sorting in emergencyEvict()
- Prevents NPE if getValue() returns null during sorting
- Defense in depth for rare concurrent removal race

§3.2 RegionBuilder: Build duration thread safety
- Snapshot both totalBuilds and totalBuildTimeNs atomically
- Prevents stale denominator in getAverageBuildTimeNs()
- Ensures correct average calculation under concurrent updates

§3.3 BuildIntegrationTest: Enhanced timeout diagnostics
- Improved error message with elapsed time and queue depth
- Better CI debugging when builds timeout
- Format: "Timeout after Xms waiting for N builds (actual: M, queue depth: Q)"

§3.4 RenderingServer: Metrics endpoint type safety
- Explicit double cast for avgBuildTimeMs calculation
- Defensive programming for division operation

Suggestions (§4.1-4.4):

§4.1 RegionCache: Enhanced Caffeine weigher warning
- Added eviction priority implications to warning message
- Clearer operational guidance when regions exceed Integer.MAX_VALUE

§4.3 Extract polling utility to shared TestUtils
- Created TestUtils class with awaitCondition() and awaitBuilds()
- Reusable across all integration tests
- BuildIntegrationTest now uses static import

§4.4 RenderingServerTest: Metrics value assertions
- Verify avgBuildTimeMs is reasonable (0.0-10000.0 range)
- Verify cache counts match (totalCount = pinnedCount + unpinnedCount)
- Catches calculation bugs in metrics endpoint

All tests passing (24 total: 19 RegionCacheTest + 4 BuildIntegrationTest + 1 testMetricsEndpoint)

+++ simulation/src/main/java/com/hellblazer/luciferase/simulation/viz/render/RenderingServer.java
@@ -303,7 +303,8 @@ public class RenderingServer implements AutoCloseable {
             builderMetrics.put("totalBuilds", regionBuilder.getTotalBuilds());
             builderMetrics.put("failedBuilds", regionBuilder.getFailedBuilds());
             builderMetrics.put("queueDepth", regionBuilder.getQueueDepth());
-            builderMetrics.put("avgBuildTimeMs", regionBuilder.getAverageBuildTimeNs() / 1_000_000.0);
+            builderMetrics.put("avgBuildTimeMs",
+                (double) regionBuilder.getAverageBuildTimeNs() / 1_000_000.0);
             metrics.put("builder", builderMetrics);
         }
